排除参数都正确，数据库卡了删除失败的情况

+++ base/src/main/java/com/rat/service/impl/AccountServiceImpl.java
@@ -136,11 +136,12 @@ public class AccountServiceImpl implements AccountService {
         else if (!userMapper.getUserByEmail(userModel.getEmail()).getPassword().equals(userModel.getPassword())) {
             return ResultTool.faild(ResultCode.USER_PASSWORD_ERROR);
         }
-        else {
+        else if (userMapper.updateUserStatusByEmail(userModel.getEmail(),USER_STATUS_DELETE) > 0){
             // 删除redis中的token
             redisUtil.deleteCache(tokenKey);
             // 将对应用户的状态设置为已删除并返回结果
             return ResultTool.success(userMapper.updateUserStatusByEmail(userModel.getEmail(),USER_STATUS_DELETE));
         }
+        return ResultTool.faild(0);
     }
 }
